--- backend/src/test/java/com/backjoongwon/cvi/post/ui/PreprocessPostControllerTest.java
@@ -1,86 +1,0 @@
-package com.backjoongwon.cvi.post.ui;
-
-import com.backjoongwon.cvi.ApiDocument;
-import com.backjoongwon.cvi.auth.domain.authorization.SocialProvider;
-import com.backjoongwon.cvi.comment.dto.CommentResponse;
-import com.backjoongwon.cvi.post.domain.Post;
-import com.backjoongwon.cvi.post.domain.VaccinationType;
-import com.backjoongwon.cvi.post.dto.PostRequest;
-import com.backjoongwon.cvi.post.dto.PostResponse;
-import com.backjoongwon.cvi.user.application.UserService;
-import com.backjoongwon.cvi.user.domain.AgeRange;
-import com.backjoongwon.cvi.user.domain.JwtTokenProvider;
-import com.backjoongwon.cvi.user.domain.User;
-import com.backjoongwon.cvi.user.dto.UserResponse;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.springframework.boot.test.mock.mockito.MockBean;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.given;
-
-@DisplayName("게시글, 좋아요, 댓글 컨트롤러 Mock 테스트 - 환경설정")
-public abstract class PreprocessPostControllerTest extends ApiDocument {
-
-    protected static final Long USER_ID = 1L;
-    protected static final Long POST_ID = 1L;
-    protected static final Long COMMENT_ID = 1L;
-    protected static final Long LIKE_ID = 1L;
-    protected static final String BEARER = "Bearer ";
-    protected static final String ACCESS_TOKEN = "{ACCESS TOKEN generated by JWT}";
-
-    @MockBean
-    protected UserService userService;
-
-    @MockBean
-    protected JwtTokenProvider jwtTokenProvider;
-
-    protected User user;
-    protected User anotherUser;
-    protected UserResponse userResponse;
-
-    protected PostRequest request;
-    protected PostResponse postResponse;
-
-    protected List<CommentResponse> commentResponses;
-
-    @BeforeEach
-    void setUp() {
-        user = User.builder()
-                .id(USER_ID)
-                .nickname("user")
-                .ageRange(AgeRange.TEENS)
-                .profileUrl("naver.com/profile")
-                .socialId("{Unique ID received from social provider}")
-                .socialProvider(SocialProvider.NAVER)
-                .build();
-        anotherUser = User.builder()
-                .id(USER_ID + 1)
-                .nickname("another_user")
-                .ageRange(AgeRange.TWENTIES)
-                .profileUrl("kakao.com/profile")
-                .socialId("{Unique ID received from social provider}")
-                .socialProvider(SocialProvider.KAKAO)
-                .build();
-        Post post = Post.builder()
-                .id(POST_ID)
-                .user(user)
-                .vaccinationType(VaccinationType.ASTRAZENECA)
-                .content("Post Content")
-                .createdAt(LocalDateTime.now())
-                .viewCount(0)
-                .build();
-
-        request = new PostRequest("글 내용", VaccinationType.PFIZER);
-
-        userResponse = UserResponse.of(user, null);
-        postResponse = PostResponse.of(post, user);
-
-        given(jwtTokenProvider.isValidToken(ACCESS_TOKEN)).willReturn(true);
-        given(jwtTokenProvider.getPayload(ACCESS_TOKEN)).willReturn(String.valueOf(user.getId()));
-        given(userService.findUserById(any(Long.class))).willReturn(user);
-    }
-}
